+++ OKX.Api/Account/Responses/OkxAccountFeeRate.cs
@@ -12,28 +12,11 @@ public record OkxAccountFeeRate
     public string Level { get; set; } = string.Empty;
 
     /// <summary>
-    /// Taker fee rate for the USDT &amp; USDⓈ &amp; Crypto trading pairs and contracts. It is the fee rate of crypto-margined contracts for FUTURES and SWAP
+    /// Fee groups.
+    /// Applicable to SPOT/MARGIN/SWAP/FUTURES/OPTION
     /// </summary>
-    [JsonProperty("taker")]
-    public decimal Taker { get; set; }
-
-    /// <summary>
-    /// Maker fee rate for the USDT &amp; USDⓈ &amp; Crypto trading pairs and contracts. It is the fee rate of crypto-margined contracts for FUTURES and SWAP
-    /// </summary>
-    [JsonProperty("maker")]
-    public decimal Maker { get; set; }
-
-    /// <summary>
-    /// Taker fee rate of USDT-margined contracts, only applicable to FUTURES/SWAP
-    /// </summary>
-    [JsonProperty("takerU")]
-    public decimal? TakerUsdtPairsAndContracts { get; set; }
-
-    /// <summary>
-    /// Maker fee rate of USDT-margined contracts, only applicable to FUTURES/SWAP
-    /// </summary>
-    [JsonProperty("makerU")]
-    public decimal? MakerUsdtPairsAndContracts { get; set; }
+    [JsonProperty("feeGroup")]
+    public List<OkxAccountFeeRateGroup> FeeGroup { get; set; } = [];
 
     /// <summary>
     /// Delivery fee rate
@@ -54,24 +37,6 @@ public record OkxAccountFeeRate
     public OkxInstrumentType InstrumentType { get; set; }
 
     /// <summary>
-    /// Taker fee rate for the USDC trading pairs(SPOT/MARGIN) and contracts(FUTURES/SWAP)
-    /// </summary>
-    [JsonProperty("takerUSDC")]
-    public decimal? TakerUsdcPairsAndContracts { get; set; }
-
-    /// <summary>
-    /// Maker fee rate for the USDC trading pairs(SPOT/MARGIN) and contracts(FUTURES/SWAP)
-    /// </summary>
-    [JsonProperty("makerUSDC")]
-    public decimal? MakerUsdcPairsAndContracts { get; set; }
-
-    /// <summary>
-    /// Trading rule types
-    /// </summary>
-    [JsonProperty("ruleType")]
-    public OkxInstrumentRuleType RuleType { get; set; }
-
-    /// <summary>
     /// Data return time, Unix timestamp format in milliseconds, e.g. 1597026383085
     /// </summary>
     [JsonProperty("ts")]
@@ -82,25 +47,18 @@ public record OkxAccountFeeRate
     /// </summary>
     [JsonIgnore]
     public DateTime Time => Timestamp.ConvertFromMilliseconds();
-
-    /// <summary>
-    /// Details of fiat fee rate
-    /// </summary>
-    [JsonProperty("fiat")]
-    public List<OkxAccountFiatFeeRate> FiatFeeRates { get; set; } = [];
 }
 
-
 /// <summary>
-/// OkxFiatFeeRate
+/// OkxAccountFeeRateGroup
 /// </summary>
-public record OkxAccountFiatFeeRate
+public record OkxAccountFeeRateGroup
 {
     /// <summary>
     /// Fiat currency
     /// </summary>
-    [JsonProperty("ccy")]
-    public string Currency { get; set; } = string.Empty;
+    [JsonProperty("groupId")]
+    public string GroupId { get; set; } = string.Empty;
     ///             
     /// <summary>
     /// Taker fee rate