+++ OKX.Api/Algo/Requests/OkxAlgoAttachedAlgoPlaceRequest.cs
@@ -19,7 +19,32 @@ public record OkxAlgoAttachedAlgoPlaceRequest
     /// </summary>
     [JsonProperty("tpTriggerPx", NullValueHandling = NullValueHandling.Ignore)]
     public string? TakeProfitTriggerPrice { get; set; }
-    
+
+    /// <summary>
+    /// Take profit trigger ratio, 0.3 represents 30%
+    /// Only one of tpTriggerPx and tpTriggerRatio can be passed
+    /// Only applicable to FUTURES and SWAP.
+    /// If the main order is a buy order, it must be greater than 0, and if the main order is a sell order, it must be bewteen -1 and 0.
+    /// </summary>
+    [JsonProperty("tpTriggerRatio", NullValueHandling = NullValueHandling.Ignore)]
+    public string? TakeProfitTriggerRatio { get; set; }
+
+    /// <summary>
+    /// Take-profit order price
+    /// For condition TP order, if you fill in this parameter, you should fill in the take-profit trigger price as well.
+    /// For limit TP order, you need to fill in this parameter, take-profit trigger needn‘t to be filled.
+    /// If the price is -1, take-profit will be executed at the market price.
+    /// </summary>
+    [JsonProperty("tpOrdPx", NullValueHandling = NullValueHandling.Ignore)]
+    public string? TakeProfitOrderPrice { get; set; }
+
+    /// <summary>
+    /// TP order kind
+    /// The default is condition
+    /// </summary>
+    [JsonProperty("tpOrdKind", NullValueHandling = NullValueHandling.Ignore)]
+    public OkxAlgoOrderKind? TakeProfitOrderKind { get; set; } = OkxAlgoOrderKind.Condition;
+
     /// <summary>
     /// Take-profit trigger price type
     /// </summary>
@@ -27,28 +52,21 @@ public record OkxAlgoAttachedAlgoPlaceRequest
     public OkxAlgoPriceType? TakeProfitTriggerPriceType { get; set; } = OkxAlgoPriceType.Last;
 
     /// <summary>
-    /// Take-profit order price
-    /// 
-    /// For condition TP order, if you fill in this parameter, you should fill in the take-profit trigger price as well.
-    /// For limit TP order, you need to fill in this parameter, take-profit trigger needn‘t to be filled.
-    /// If the price is -1, take-profit will be executed at the market price.
-    /// </summary>
-    [JsonProperty("tpOrdPx", NullValueHandling = NullValueHandling.Ignore)]
-    public string? TakeProfitOrderPrice { get; set; }
-    
-    /// <summary>
     /// Stop-loss trigger price
     /// If you fill in this parameter, you should fill in the stop-loss order price.
     /// </summary>
     [JsonProperty("slTriggerPx", NullValueHandling = NullValueHandling.Ignore)]
     public string? StopLossTriggerPrice { get; set; }
-    
+
     /// <summary>
-    /// Stop-loss trigger price type
+    /// Stop profit trigger ratio, 0.3 represents 30%
+    /// Only one of slTriggerPx and slTriggerRatio can be passed
+    /// Only applicable to FUTURES and SWAP.
+    /// If the main order is a buy order, it should be bewteen 0 and 1, and if the main order is a sell order, it must be greater than 0.
     /// </summary>
-    [JsonProperty("slTriggerPxType", NullValueHandling = NullValueHandling.Ignore)]
-    public OkxAlgoPriceType? StopLossTriggerPriceType { get; set; } = OkxAlgoPriceType.Last;
-    
+    [JsonProperty("slTriggerRatio", NullValueHandling = NullValueHandling.Ignore)]
+    public string? StopLossTriggerRatio { get; set; }
+
     /// <summary>
     /// Stop-loss order price
     /// If you fill in this parameter, you should fill in the stop-loss trigger price.
@@ -56,4 +74,24 @@ public record OkxAlgoAttachedAlgoPlaceRequest
     /// </summary>
     [JsonProperty("slOrdPx", NullValueHandling = NullValueHandling.Ignore)]
     public string? StopLossOrderPrice { get; set; }
+
+    /// <summary>
+    /// Stop-loss trigger price type
+    /// </summary>
+    [JsonProperty("slTriggerPxType", NullValueHandling = NullValueHandling.Ignore)]
+    public OkxAlgoPriceType? StopLossTriggerPriceType { get; set; } = OkxAlgoPriceType.Last;
+
+    /// <summary>
+    /// Size. Only applicable to TP order of split TPs, and it is required for TP order of split TPs
+    /// </summary>
+    [JsonProperty("sz", NullValueHandling = NullValueHandling.Ignore)]
+    public string? Size { get; set; }
+
+    /// <summary>
+    /// Whether to enable Cost-price SL. Only applicable to SL order of split TPs. Whether slTriggerPx will move to avgPx when the first TP order is triggered
+    /// 0: disable, the default value
+    /// 1: Enable
+    /// </summary>
+    [JsonProperty("amendPxOnTriggerType", NullValueHandling = NullValueHandling.Ignore)]
+    public string? AmendPriceOnTriggerType { get; set; }
 }