+++ OKX.Api/Public/Responses/OkxPublicInstrument.cs
@@ -32,6 +32,50 @@ public record OkxPublicInstrument
     public string? Underlying { get; set; } = string.Empty;
 
     /// <summary>
+    /// Instrument trading fee group ID
+    /// Spot:
+    /// 1: Spot USDT
+    /// 2: Spot USDC &amp; Crypto
+    /// 3: Spot TRY
+    /// 4: Spot EUR
+    /// 5: Spot BRL
+    /// 7: Spot AED
+    /// 8: Spot AUD
+    /// 9: Spot USD
+    /// 10: Spot SGD
+    /// 11: Spot zero
+    /// 12: Spot group one
+    /// 13: Spot group two
+    /// 14: Spot group three
+    /// 15: Spot special rule
+    /// 
+    /// Expiry futures:
+    /// 1: Expiry futures crypto-margined
+    /// 2: Expiry futures USDT-margined
+    /// 3: Expiry futures USDC-margined
+    /// 4: Expiry futures premarket
+    /// 5: Expiry futures group one
+    /// 6: Expiry futures group two
+    /// 
+    /// Perpetual futures:
+    /// 1: Perpetual futures crypto-margined
+    /// 2: Perpetual futures USDT-margined
+    /// 3: Perpetual futures USDC-margined
+    /// 4: Perpetual futures group one
+    /// 5: Perpetual futures group two
+    /// 
+    /// Options:
+    /// 1: Options crypto-margined
+    /// 2: Options USDC-margined
+    /// 
+    /// instType and groupId should be used together to determine a trading fee group.Users should use this endpoint together with fee rates endpoint to get the trading fee of a specific symbol.
+    /// 
+    /// Some enum values may not apply to you; the actual return values shall prevail.
+    /// </summary>
+    [JsonProperty("groupId")]
+    public int? GroupId { get; set; }
+
+    /// <summary>
     /// Instrument family
     /// </summary>
     [JsonProperty("instFamily")]
@@ -205,24 +249,6 @@ public record OkxPublicInstrument
     public OkxInstrumentRuleType? RuleType { get; set; }
 
     /// <summary>
-    /// Maximum position value (USD) for this instrument at the user level, based on the notional value of all same-direction open positions and resting orders. The effective user limit is max(posLmtAmt, oiUSD × posLmtPct). Applicable to SWAP/FUTURES.
-    /// </summary>
-    [JsonProperty("posLmtAmt")]
-    public decimal? MaximumPositionValue { get; set; }
-
-    /// <summary>
-    /// Maximum position ratio (e.g., 30 for 30%) a user may hold relative to the platform’s current total position value. The effective user limit is max(posLmtAmt, oiUSD × posLmtPct). Applicable to SWAP/FUTURES.
-    /// </summary>
-    [JsonProperty("posLmtPct")]
-    public decimal? MaximumPositionRatio { get; set; }
-
-    /// <summary>
-    /// Platform-wide maximum position value (USD) for this instrument. If the global position limit switch is enabled and platform total open interest reaches or exceeds this value, all users’ new opening orders for this instrument are rejected; otherwise, orders pass.
-    /// </summary>
-    [JsonProperty("maxPlatOILmt")]
-    public decimal? PlatformWideMaximumPositionValue { get; set; }
-
-    /// <summary>
     /// Maximal limit order size
     /// </summary>
     [JsonProperty("maxLmtSz")]
@@ -282,4 +308,61 @@ public record OkxPublicInstrument
     /// </summary>
     [JsonProperty("tradeQuoteCcyList")]
     public List<string>? TradeQuoteCurrencyList { get; set; } = [];
+
+    /// <summary>
+    /// Maximum position value (USD) for this instrument at the user level, based on the notional value of all same-direction open positions and resting orders. The effective user limit is max(posLmtAmt, oiUSD × posLmtPct). Applicable to SWAP/FUTURES.
+    /// </summary>
+    [JsonProperty("posLmtAmt")]
+    public decimal? MaximumPositionValue { get; set; }
+
+    /// <summary>
+    /// Maximum position ratio (e.g., 30 for 30%) a user may hold relative to the platform’s current total position value. The effective user limit is max(posLmtAmt, oiUSD × posLmtPct). Applicable to SWAP/FUTURES.
+    /// </summary>
+    [JsonProperty("posLmtPct")]
+    public decimal? MaximumPositionRatio { get; set; }
+
+    /// <summary>
+    /// Platform-wide maximum position value (USD) for this instrument. If the global position limit switch is enabled and platform total open interest reaches or exceeds this value, all users’ new opening orders for this instrument are rejected; otherwise, orders pass.
+    /// </summary>
+    [JsonProperty("maxPlatOILmt")]
+    public decimal? PlatformWideMaximumPositionValue { get; set; }
+
+    /// <summary>
+    /// Upcoming changes. It is [] when there is no upcoming change.
+    /// </summary>
+    [JsonProperty("upcChg")]
+    public List<OkxPublicInstrumentUpcomingChange>? UpcomingChanges { get; set; }
+}
+
+/// <summary>
+/// OKX Instrument
+/// </summary>
+public record OkxPublicInstrumentUpcomingChange
+{
+    /// <summary>
+    /// The parameter name to be updated.
+    /// tickSz
+    /// minSz
+    /// maxMktSz
+    /// </summary>
+    [JsonProperty("param")]
+    public string Parameter { get; set; } = "";
+
+    /// <summary>
+    /// The parameter value that will replace the current one.
+    /// </summary>
+    [JsonProperty("newValue")]
+    public string NewValue { get; set; } = "";
+
+    /// <summary>
+    /// Effective timestamp
+    /// </summary>
+    [JsonProperty("effTime")]
+    public long EffectiveTimestamp { get; set; }
+
+    /// <summary>
+    /// Effective time
+    /// </summary>
+    [JsonIgnore]
+    public DateTime EffectiveTime => EffectiveTimestamp.ConvertFromMilliseconds();
 }